Try to get all results from ES query
This is not robust, as it has hardcoded upper bounds (1000 entries
for each glossary, and 120 glossaries in total), but it will help
us check against other results and have a fuller entry list. Once
we start indexing entries as individual documents (rather than
nesting them under glossaries), this will be easier to implement
in a nice way. The pagination interface doesn't seem to work for
nested queries, unfortunately.