Explicitly return response, not raw JSON
Using jsonify is clearer about returning an HTTP response. The
end result is the same (Flask wraps the return value in a response
anyway), but it is pretty-printed and can be customised later on
(eg by specifying the response code). In particular, note that the
message in the response returned is a list of JSON objects, as
before (which is what the Angular frontend expects).