Add missing data to test glossary
The test glossary contains the elx glossary plus an entry from the
xhu glossary. The summary and instance id corresponding to the xhu
entry had been left out by accident. This didn't make a difference
because so far we were completely ignoring those fields. Now that
we are linking the instances, that field needs to be added, and I'm
including the summary just so it looks like a complete "authentic"
glossary, even if we are likely to ignore it. Note that this now
includes all relevant instances, even ones which are not referenced
by the entry but are period-specific variants of the referenced
ones, i.e. xhu.r00006.p.s000 and xhu.r00007.p.s000.