Remove "from" option in favour of starting entry
There are now two main ways of calling the search_all endpoint:
We can give an "after" argument, to specify a starting point (for
pagination), or we can retrieve entries from the very start of the
results (by not specifying "after"). In either way, we can pass a
field to sort on ("sort_by", defaults to "cf") and a direction
("dir", defaults to ascending). We can also pass a "count" argument
to specify the maximum number of results to retrieve. If "count" is
not given, then we either get back 10 hits by default (if using
"after"), or all results (if not passing "after").

This last point is a bit confusing, in that the default number of
hits is different for the two "modes". This might be changed in the
future to make it so "count" is required if "after" is also given.