Specify the mapping for a local-aware field
Instead of just using the cf.keyword field, which follows an
English-based sorting (and therefore sends all non-standard letter
variants to the end), we create the new field cf.sort, which will
(by default) follow the DUCET collation, a good choice for
languages which don't have a standardised/implemented collation.
This will e.g. sort variants of s right after the base s, and give
an acceptable sorting.
This also means we definitely need to keep the TYPE_NAME, as it's
used in the mapping definition.