Include sort score in server response
The results returned now include a "sort" field, which contains
the sorting score of the corresponding result. This could be used
to do pagination, if we need to know the last retrieved score.
Doing this involved a minor change in the single-field search
(which is hardly used, and should not affect its functionality).

Also removed a TODO comment which no longer applied (I believe all
entries should now have the required fields, after we switched to
the new individual-entry indexing).