Use a tie-breaker field when sorting
Two or more entries can have the same value for a particular field.
This causes problems when we are sorting on that field, as we may
miss some results when using the `after` parameter (for example,
after=5 will return entries with icount >= 6, even if we haven't
seen all entries with icount = 5 yet). We therefore need to include
a tiebreaker field that uniquely identifies each entry, to be used
as a fallback when there are multiple entries with the given value
in the sorting field. See also:
https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-search-after.html