Return partial matches
Up to now we only searched for exact matches, which was never the
intention! With this change, searching for e.g. "water" will also
return entries containing "waters" and "waterclock".

Apparently this only works with text, not keyword fields
(see https://stackoverflow.com/questions/46696190)
but this covers all the fields we are searching for
(we do have some keyword fields, but those are only used for
sorting, not matching).