Catch error from ingesting Akkadian glossary
One of the entries in the Akkadian glossary we are using refers to
instance ".r000001", which does not appear to exist, causing an
errror when indexing the glossary. We should investigate why this
is but, in the meantime, this just skips any entries with missing
instances and gives a message during ingestion.