Fix encoding!
The subprocess call returns bytes, not a string, which made the
ensuing JSON load fail. We just need to decode the bytes to a
string object. Another way is to pass an `encoding="utf8"` argument
to `subprocess.run`, but this was only introduced in Python 3.6
(the Azure test site has 3.5).
Tests pass with this change.