Don't search headword, but keep it in results
The headword field poses some problems for the new analyzer we
are planning.

The analyzer will contain a filter to replace non-ASCII characters
with ASCII "synonyms". Some of these synonyms contain characters
that the standard tokenizer strips out, therefore the filter cannot
be used at the same time as that tokenizer.
The headword field is unique in that it both contains non-ASCII
characters and needs to be tokenized (to separate the guideword
from the cuneiform and part of speech). This means that we would
need some more complex way of indexing it.
Fortunately, it contains the same information as gw and cf (as well
as the part of speech, which we don't do anything with at the
moment), so we don't lose anything by removing it from the search.