Require ElasticSearch 6, not 7
Version 7 of ElasticSearch has introduced some breaking changes,
particularly the removal of document types. This will not be hard
to adapt to (we only have one document type in our index anyway),
but newer versions of the elasticsearch_dsl package have changed
their API to follow this. As we mostly use ES 5/6 locally and in
the Oracc server, for now we should stick to the previous version
of the package.