Fix Index creation for previous library versions
Earlier versions of elasticsearch_dsl set a default document type
when creating an index. This leads to an error when we try to add
the mapping for our own document type ("entry"), because an index
can only have one type. This fix avoids this by passing the
intended document type from the start.
(This problems appears in versions including 6.2.1, but has been
fixed by 6.4.0)